rag-pipeline/requirements.txt
fastapi
uvicorn
PyPDF2
camelot-py[cv]
sentence-transformers
faiss-cpu
openai
pandas
